Attributes on Javascript and css
Just wanted to add an example of attributes on js. Took me a while to figure it out. I was used to jade/pug. In that syntax you do 
```jade
  script(attr="val").
    var name=""
```
so I though slang would have the colon at the end, eg `javascript type="text/javascript" :`, I was obviously wrong.